==> 6-FC_lme_masterscheme.r <==
# Using the linear mixed model approach for fungal competition.
# August 2019

# I should build one model per output parameter that interests me.
# What are the most interesting options?

# 1) Total C-13 enrichment of mycorrhizas
# 2) Carbon allocation ratio (Tt only)
# 3) C for N exchange rate by fungal identity and competitor identity

# I should also use the mixed model framework
# for my C-for-N proof-of-concept analyses, really.

# 1) Does mycorrhiza C track hyphal C? (all mycos with corresponding rhizomorph)
# 2) Does mycorrhiza N track root N? (only labeled side of root systems)
# 3) Do mycorrhizas that had more N enrichment also get more C? (only labeled side of root systems)
